+++ src/Control/Lens/Internal/ByteString.hs
@@ -9,6 +9,9 @@
 #ifndef MIN_VERSION_base
 #define MIN_VERSION_base(x,y,z) 1
 #endif
+
+{-# OPTIONS_GHC -fno-warn-deprecations #-} -- for inlinePerformIO
+
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Data.ByteString.Strict.Lens
@@ -42,7 +45,10 @@ import Data.Int (Int64)
 import Data.Word (Word8)
 import Foreign.Ptr
 import Foreign.Storable
-#if MIN_VERSION_base(4,4,0)
+#if MIN_VERSION_base(4,8,0)
+import Foreign.ForeignPtr
+import Foreign.ForeignPtr.Unsafe
+#elif MIN_VERSION_base(4,4,0)
 import Foreign.ForeignPtr.Safe
 import Foreign.ForeignPtr.Unsafe
 #else